+++ code/header/block/first_order_block.h
@@ -24,7 +24,7 @@ class FIRST_ORDER_BLOCK : public BLOCK
     private:
         void integrate();
         void update();
-        double K, T;
+        float K, T;
 
         double one_over_t, one_over_h, one_over_k, t_over_h, h_plus_2t, one_over_h_plus_2t, h_minus_2t;
 };
