+++ code/header/device/line.h
@@ -112,15 +112,15 @@ class LINE : public DEVICE
         BUS* sending_side_busptr, *receiving_side_busptr;
         string   identifier;
         bool sending_side_breaker_status, receiving_side_breaker_status;
-        complex<double> line_z_12, line_z_0,
-                        line_y_12, line_y_0,
-                        shunt_y_12_sending_side, shunt_y_0_sending_side,
-                        shunt_y_12_receiving_side, shunt_y_0_receiving_side;
+        complex<float> line_z_12, line_z_0,
+                       line_y_12, line_y_0,
+                       shunt_y_12_sending_side, shunt_y_0_sending_side,
+                       shunt_y_12_receiving_side, shunt_y_0_receiving_side;
         RATING rating;
         unsigned int meter_end_bus;
-        double length;
+        float length;
 
-        map<double,FAULT> faults;
+        map<float,FAULT> faults;
 };
 
 #endif // LINE_H
