+++ code/header/model/sg_models/compensator_model/PSASPVC.h
@@ -44,7 +44,7 @@ class PSASPVC: public COMPENSATOR_MODEL
     private:
         void copy_from_const_model(const PSASPVC& model);
 
-        double Xc;
+        float Xc;
         double sin_phi;
 };
