+++ code/header/model/sg_models/turbine_governor_model/IEEEG1SB.h
@@ -105,23 +105,23 @@ class IEEEG1SB : public TURBINE_GOVERNOR_MODEL
         void copy_from_const_model(const IEEEG1SB& model);
 
         LEAD_LAG_BLOCK droop;
-        double Tservo;
-        double Uo, Uc;
+        float Tservo;
+        float Uo, Uc;
         INTEGRAL_BLOCK servo_motor;
         FIRST_ORDER_BLOCK delayer1;
-        double K1, K2;
+        float K1, K2;
         FIRST_ORDER_BLOCK delayer2;
-        double K3, K4;
+        float K3, K4;
         FIRST_ORDER_BLOCK delayer3;
-        double K5, K6;
+        float K5, K6;
         FIRST_ORDER_BLOCK delayer4;
-        double K7, K8;
+        float K7, K8;
         PI_BLOCK boiler_PI;
         FIRST_ORDER_BLOCK fuel_producer;
         FIRST_ORDER_BLOCK fuel_delayer1;
         FIRST_ORDER_BLOCK fuel_delayer2;
         INTEGRAL_BLOCK drum;
-        double Kb;
+        float Kb;
 
 };
