+++ code/header/model/wtg_models/wt_aerodynamic_model/wt_aerodynamic_model.h
@@ -131,13 +131,13 @@ class WT_AERODYNAMIC_MODEL : public WTG_MODEL
         double get_current_pelec_including_loss_per_turbine_in_MW() const;
         double get_current_turbine_speed_reference_without_limit_in_rad_per_s() const;
         // turbine nominals
-        double nominal_wind_speed_in_mps;
+        float nominal_wind_speed_in_mps;
         double turbine_blade_radius_in_m;
-        double nominal_air_density_in_kgpm3;
+        float nominal_air_density_in_kgpm3;
 
         unsigned int number_of_pole_pairs;
         double generator_to_turbine_gear_ratio;
-        double gear_efficiency;
+        float gear_efficiency;
 
         // operation parameter
         double initial_wind_speed_in_mps;
@@ -150,7 +150,7 @@ class WT_AERODYNAMIC_MODEL : public WTG_MODEL
         double turbine_speed_in_rad_per_s;
         WTG_TURBINE_SPEED_MODE speed_mode;
 
-        double max_steady_state_turbine_speed_in_pu, min_steady_state_turbine_speed_in_pu;
+        float max_steady_state_turbine_speed_in_pu, min_steady_state_turbine_speed_in_pu;
 
         double cpmax_at_zero_pitch;
         double current_pitch_angle_in_deg, current_lambda_at_cpmax;
