+++ code/header/model/wtg_models/wt_pitch_model/wt_pitch_model.h
@@ -66,9 +66,9 @@ class WT_PITCH_MODEL : public WTG_MODEL
     private:
         void common_constructor();
 
-        double frequency_deadband_upper_in_pu, frequency_deadband_lower_in_pu;
+        float frequency_deadband_upper_in_pu, frequency_deadband_lower_in_pu;
         bool hold_wtg_speed_flag;
-        double const_wtg_speed_reference;
+        float const_wtg_speed_reference;
 };
 
 #endif // WT_PITCH_MODEL_H
